# apps/frontend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/workspaces/{wsId}/supports": {
      "get": {
        "operationId": "SupportsController_findAll",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teamId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "themeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minScore",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List supports in workspace",
        "tags": [
          "Supports"
        ]
      },
      "post": {
        "operationId": "SupportsController_create",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a support request",
        "tags": [
          "Supports"
        ]
      }
    },
    "/api/workspaces/{wsId}/supports/{id}": {
      "get": {
        "operationId": "SupportsController_findOne",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get support detail",
        "tags": [
          "Supports"
        ]
      },
      "patch": {
        "operationId": "SupportsController_update",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a support request",
        "tags": [
          "Supports"
        ]
      },
      "delete": {
        "operationId": "SupportsController_remove",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a support request",
        "tags": [
          "Supports"
        ]
      }
    },
...
    "/api/workspaces/{wsId}/test-automation/junit": {
      "post": {
        "operationId": "TestAutomationController_importJunit",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import CI test results from JUnit XML → auto-match/create test cases → create test runs",
        "tags": [
          "Test Automation"
        ]
      }
    },
...
    "/api/workspaces/{wsId}/sprints": {
      "get": {
        "operationId": "SprintsController_list",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "releaseId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List sprints (optionally by release)",
        "tags": [
          "sprints"
        ]
      },
      "post": {
        "operationId": "SprintsController_create",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSprintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create sprint",
        "tags": [
          "sprints"
        ]
      }
    },
    "/api/workspaces/{wsId}/sprints/{id}/stats": {
      "get": {
        "operationId": "SprintsController_stats",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Sprint stats (progress / capacity)",
        "tags": [
          "sprints"
        ]
      }
    },
    "/api/workspaces/{wsId}/sprints/{id}": {
      "patch": {
        "operationId": "SprintsController_update",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSprintDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update sprint",
        "tags": [
          "sprints"
        ]
      },
      "delete": {
        "operationId": "SprintsController_remove",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete sprint (blocked if stories attached)",
        "tags": [
          "sprints"
        ]
      }
    },
...
    "/api/telemetry/updates/check": {
      "get": {
        "operationId": "TelemetryController_checkUpdate",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "edition",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Check latest version; instance heartbeat recorded as side effect",
        "tags": [
          "Telemetry"
        ]
      }
    },
    "/api/telemetry/instances": {
      "get": {
        "operationId": "TelemetryController_listInstances",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List active instances (admin)",
        "tags": [
          "Telemetry"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pinecone API",
    "description": "Aha!-style requirements management system.\n\nI10 API 治理（2026-08-18 P2）：全局前缀 /api；认证 = Bearer JWT（Web）或 API Token（pc_tok_ 前缀，SHA-256 存储）；全局限流 100 req/min/IP（Throttler）；Swagger 仅非生产暴露。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSupportDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_REVIEW",
              "CLOSED"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "SUPPORT_REQUEST",
              "DEFECT"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "MAJOR",
              "MINOR",
              "TRIVIAL"
            ],
            "description": "缺陷严重度（type=DEFECT 时有效）"
          },
          "rootCause": {
            "type": "string",
            "description": "缺陷根因（自由文本，type=DEFECT 时选填）"
          },
          "discoveryPhase": {
            "type": "string",
            "enum": [
              "TEST",
              "PRODUCTION",
              "CUSTOMER"
            ],
            "description": "缺陷发现阶段（type=DEFECT 时选填，逃逸率口径）"
          },
          "assigneeId": {
            "type": "string"
          },
          "assigneeName": {
            "type": "string"
          },
          "releaseId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateSupportDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_REVIEW",
              "CLOSED"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "SUPPORT_REQUEST",
              "DEFECT"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "MAJOR",
              "MINOR",
              "TRIVIAL"
            ],
            "description": "缺陷严重度（type=DEFECT 时有效）"
          },
          "rootCause": {
            "type": "string",
            "description": "缺陷根因（自由文本，type=DEFECT 时选填）"
          },
          "discoveryPhase": {
            "type": "string",
            "enum": [
              "TEST",
              "PRODUCTION",
              "CUSTOMER"
            ],
            "description": "缺陷发现阶段（type=DEFECT 时选填，逃逸率口径）"
          },
          "assigneeId": {
            "type": "string"
          },
          "assigneeName": {
            "type": "string"
          },
          "releaseId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateSprintDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSprintDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}